Export indexts updated
