--- src/lib/components/Table/DataTable.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-	// ...
-</script>
-
-<h2>(working!)</h2>
