+++ src/routes/(inner)/guides/styling/+page.svelte
@@ -89,6 +89,7 @@
 		head: ['Stylesheet', 'Documentation', 'View Source'],
 		// prettier-ignore
 		body: [
+			// Tailwind Elements
 			[
 				'<code>elements/badges.css</code>',
 				'<a href="/elements/badges">badges</a>',
@@ -119,6 +120,12 @@
 				'<a href="/elements/placeholders">placeholders</a>',
 				`<a href="${ghLibPathMaster}/styles/elements/placeholders.css" target="_blank" rel="noreferrer">placeholders.css</a>`,
 			],
+			[
+				'<code>elements/tables.css</code>',
+				'<a href="/utilities/tables">tables</a>',
+				`<a href="${ghLibPathMaster}/styles/elements/tables.css" target="_blank" rel="noreferrer">tables.css</a>`,
+			],
+			// Utilities
 			[
 				'<code>elements/menus.css</code>',
 				'<a href="/utilities/menus">menus</a>',
@@ -184,10 +191,14 @@ import '@brainandbones/skeleton/styles/${$storeStylesheets === 'recommended' ? '
 			<Tab value="advanced">Advanced</Tab>
 		</TabGroup>
 		{#if $storeStylesheets === 'recommended'}
-			<p>We recommend <code>all.css</code> for most users. This includes everything required for Skeleton, with all imports in the correct order.</p>
+			<p>
+				We recommend <code>all.css</code> for most users. This includes everything required for Skeleton, with all imports in the correct order.
+			</p>
 			<Table source={tableStyleAll} />
 		{:else if $storeStylesheets === 'advanced'}
-			<p>For advanced users ONLY. Follow the instruction below to import each stylesheet individually. Please ensure you use the order shown.</p>
+			<p>
+				For advanced users ONLY. Follow the instruction below to import each stylesheet individually. Please ensure you use the order shown.
+			</p>
 			<Table source={tableStylesUseful} />
 			<!-- Elements -->
 			<TabGroup selected={storeStylesheetElements}>
@@ -197,7 +208,9 @@ import '@brainandbones/skeleton/styles/${$storeStylesheets === 'recommended' ? '
 			{#if $storeStylesheetElements === 'combined'}
 				<Table source={tableStyleElementsAll} />
 			{:else if $storeStylesheetElements === 'seperate'}
-				<p>Import only the Tailwind Elements you are using. Make sure you set the path to <code>.../styles/elements/*.css</code> as shown.</p>
+				<p>
+					Import only the Tailwind Elements you are using. Make sure you set the path to <code>.../styles/elements/*.css</code> as shown.
+				</p>
 				<Table source={tableStyleElements} />
 			{/if}
 		{/if}
@@ -208,7 +221,10 @@ import '@brainandbones/skeleton/styles/${$storeStylesheets === 'recommended' ? '
 	<!-- Global Styles -->
 	<section class="space-y-4">
 		<h2>Global Stylesheet</h2>
-		<p>The location of your app's global stylesheet differs per framework. SvelteKit and Vite users are required to make one quick modification.</p>
+		<p>
+			The location of your app's global stylesheet differs per framework. SvelteKit and Vite users are required to make one quick
+			modification.
+		</p>
 
 		<TabGroup selected={storeFramework}>
 			<Tab value="sveltekit">SvelteKit</Tab>
@@ -224,7 +240,8 @@ import '@brainandbones/skeleton/styles/${$storeStylesheets === 'recommended' ? '
 			<!-- Framework: Astro -->
 		{:else if $storeFramework === 'astro'}
 			<p>
-				Astro does not create this by default, but we recommend <code>/src/styles/global.css</code>. <a href="/guides/frameworks/astro">See our Astro walkthrough</a> for assistance.
+				Astro does not create this by default, but we recommend <code>/src/styles/global.css</code>.
+				<a href="/guides/frameworks/astro">See our Astro walkthrough</a> for assistance.
 			</p>
 		{/if}
 		<Alert>
@@ -234,8 +251,9 @@ import '@brainandbones/skeleton/styles/${$storeStylesheets === 'recommended' ? '
 		<!-- Directives -->
 		<p>
 			When using <strong>Svelte-Add</strong> this creates
-			<a href="https://tailwindcss.com/docs/functions-and-directives" target="_blank" rel="noreferrer">@tailwind directives</a> in your global stylesheet. This is not recommended as Skeleton handles this
-			as part of the stylesheet imports above. If present, remove the following from your global stylesheet.
+			<a href="https://tailwindcss.com/docs/functions-and-directives" target="_blank" rel="noreferrer">@tailwind directives</a> in your global
+			stylesheet. This is not recommended as Skeleton handles this as part of the stylesheet imports above. If present, remove the following
+			from your global stylesheet.
 		</p>
 		<CodeBlock
 			language="css"
@@ -262,7 +280,10 @@ import '@brainandbones/skeleton/styles/${$storeStylesheets === 'recommended' ? '
 	<section class="space-y-4">
 		<div class="space-y-4">
 			<h2>Styling Components</h2>
-			<p>Skeleton components automatically adapt to your theme. However, you may wish to customize your components and elements. View instruction for this below.</p>
+			<p>
+				Skeleton components automatically adapt to your theme. However, you may wish to customize your components and elements. View
+				instruction for this below.
+			</p>
 		</div>
 		<div class="card card-body !bg-accent-500/5">
 			<AccordionGroup>
@@ -270,8 +291,8 @@ import '@brainandbones/skeleton/styles/${$storeStylesheets === 'recommended' ? '
 					<svelte:fragment slot="summary"><h3>Via Component Props</h3></svelte:fragment>
 					<div slot="content" class="space-y-4">
 						<p>
-							This is the recommended way to style most components. Provide style "props" (aka properties) that allow you to provide utility classes to override styles. See a full list of available
-							settings under the "Props" tab for each component's documentation.
+							This is the recommended way to style most components. Provide style "props" (aka properties) that allow you to provide utility
+							classes to override styles. See a full list of available settings under the "Props" tab for each component's documentation.
 						</p>
 						<CodeBlock language="html" code={`<Tab background="bg-accent-500" style="bg-yellow-500">Prop Customized</Tab>`} />
 					</div>
@@ -280,8 +301,8 @@ import '@brainandbones/skeleton/styles/${$storeStylesheets === 'recommended' ? '
 					<svelte:fragment slot="summary"><h3>Via the Class Attribute</h3></svelte:fragment>
 					<div slot="content" class="space-y-4">
 						<p>
-							If a particular style prop is not provided, you can still provide arbitrary utility classes via the standard <code>class</code> attribute on any component. Note these classes are applied
-							to the parent element in the component template.
+							If a particular style prop is not provided, you can still provide arbitrary utility classes via the standard <code>class</code
+							> attribute on any component. Note these classes are applied to the parent element in the component template.
 						</p>
 						<CodeBlock language="html" code={`<Tab class="text-3xl px-10 py-5">Big</Tab>`} />
 					</div>
@@ -303,15 +324,19 @@ import '@brainandbones/skeleton/styles/${$storeStylesheets === 'recommended' ? '
 					<svelte:fragment slot="summary"><h3>Global Styles</h3></svelte:fragment>
 					<div slot="content" class="space-y-4">
 						<p>
-							Tailwind Elements and Svelte Components contain a unique selector classes, such as <code>.crumb-separator</code> for the Breadcrumb component seperator element. Use these to target global
-							style overrides.
+							Tailwind Elements and Svelte Components contain a unique selector classes, such as <code>.crumb-separator</code> for the Breadcrumb
+							component seperator element. Use these to target global style overrides.
 						</p>
-						<CodeBlock language="html" code={`<!--  The first class is the "selector" class -->\n<div class="crumb-separator ...">&rarr;</div>`} />
+						<CodeBlock
+							language="html"
+							code={`<!--  The first class is the "selector" class -->\n<div class="crumb-separator ...">&rarr;</div>`}
+						/>
 						<p>Add the following to your global stylesheet to override the seperator text color:</p>
 						<CodeBlock language="css" code={`.crumb-separator { @apply !text-red-500; }`} />
 						<p>
 							Note that in some cases you may need to use <code>!</code>
-							<a href="https://tailwindcss.com/docs/configuration#important-modifier" target="_blank" rel="noreferrer">important</a> to give the class precedence.
+							<a href="https://tailwindcss.com/docs/configuration#important-modifier" target="_blank" rel="noreferrer">important</a> to give
+							the class precedence.
 						</p>
 					</div>
 				</AccordionItem>
